Updating Config file to reflect new default behaviour
The standard behaviour has changed: We now assume that people
downloading the code are doing so because they want to run it on a
Triggertrap Shield for Arduino. People who are developing on the TTv1,
need to comment out the #define TT_SHIELD statement in the config file.

This update amends the in-line help to reflect this change.